+++ include/overture.h
@@ -2,23 +2,24 @@
 #define _OVERTURE_H_
 
 #include "process.h"
+#include "common.h"
 
 typedef struct {
-    int port;
-    int debug;
-    int timeout;
+    uint16_t port;
+    uint8_t debug; // bool value
+    uint32_t timeout;
     char *ttl_file;
     char *host_file;
-    int foreign_port;
-    int domestic_port;
-    int **reject_type;
+    uint16_t foreign_port;
+    uint16_t domestic_port;
+    uint32_t **reject_type;
     char *foreign_ip_file;
     char *domestic_ip_file;
     char *foreign_domain_file;
     char *domestic_domain_file;
 } overture;
 
-overture* overture_init(int port);
+overture* overture_init(uint16_t port);
 process* overture_load(overture *info, const char *file);
 
 #endif
